requirements.txt
openai
python-dotenv
paramiko
duckduckgo_search